Creacion de DatailViewModel y HomeViewModel
Separación de responsabilidades en ViewModels según MVVM: se separaron las responsabilidades de MoviesViewModel en HomeViewModel y DetailViewModel. Se ajustó la inyección de dependencias.
