+++ composeApp/src/desktopMain/kotlin/edu/dyds/movies/presentation/Navigation.kt
@@ -10,7 +10,8 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
-import edu.dyds.movies.di.MoviesDependencyInjector.getMoviesViewModel
+import edu.dyds.movies.di.MoviesDependencyInjector.getDetailViewModel
+import edu.dyds.movies.di.MoviesDependencyInjector.getHomeViewModel
 import edu.dyds.movies.presentation.detail.DetailScreen
 import edu.dyds.movies.presentation.home.HomeScreen
 
@@ -33,7 +34,7 @@ fun Navigation() {
 private fun NavGraphBuilder.homeDestination(navController: NavHostController) {
     composable(HOME) {
         HomeScreen(
-            viewModel = getMoviesViewModel(),
+            homeViewModel = getHomeViewModel(),
             onGoodMovieClick = {
                 navController.navigate("$DETAIL/${it.id}")
             }
@@ -49,7 +50,7 @@ private fun NavGraphBuilder.detailDestination(navController: NavHostController) 
         val movieId = backstackEntry.arguments?.getInt(MOVIE_ID)
 
         movieId?.let {
-            DetailScreen(getMoviesViewModel(), it, onBack = { navController.popBackStack() })
+            DetailScreen(getDetailViewModel(), it, onBack = { navController.popBackStack() })
         }
     }
 }
